generic-mqtt-service/requirements.txt
viam-sdk
paho-mqtt